feat: 헤더 UI 대략 완료
