+++ .eslintrc.json
@@ -14,7 +14,8 @@
   "rules": {
     "unused-imports/no-unused-imports": "error",
     "simple-import-sort/imports": "error",
-    "@typescript-eslint/semi": ["error", "never"]
+    "@typescript-eslint/semi": ["error", "never"],
+    "@typescript-eslint/no-use-before-define": "off"
   },
   "ignorePatterns": "*.js"
 }
