+++ app/community/page.tsx
@@ -1,0 +1,7 @@
+export default function Community() {
+  return (
+    <div>
+      <h2>커뮤니티</h2>
+    </div>
+  )
+}
