+++ app/layout/Header/components/ControlPanel.tsx
@@ -1,0 +1,12 @@
+export default function ControlPanel() {
+  return (
+    <div>
+      <button
+        type="button"
+        className="border border-stone-800 text-stone-600 px-2 py-1 rounded-sm hover:outline hover:outline-1 hover:outline-stone-400"
+      >
+        로그인
+      </button>
+    </div>
+  )
+}
