+++ app/layout/Header/components/Logo.tsx
@@ -1,0 +1,3 @@
+export default function Logo() {
+  return <h2 className="text-2xl  tracking-widest font-bold text-orange-700">SAJA</h2>
+}
