+++ app/layout/Header/components/TopSection.tsx
@@ -1,13 +1,12 @@
+import ControlPanel from './ControlPanel'
+import Logo from './Logo'
+
 export default function TopSection() {
   return (
-    <section className="w-full h-[20px]">
+    <section className="w-full h-[52px]">
       <div className="max-w-6xl h-full m-auto px-10 flex items-center justify-between">
-        <div>
-          로그
-        </div>
-        <div>
-          <button type="button">로그인</button>
-        </div>
+        <Logo />
+        <ControlPanel />
       </div>
     </section>
   )
