+++ app/layout/Header/constants/navInfo.ts
@@ -4,16 +4,12 @@ const navInfo = [
     link: '/mentoring',
   },
   {
-    name: '블로그',
-    link: '/blog',
-  },
-  {
     name: '커뮤니티',
     link: '/community',
   },
   {
-    name: '레쥬메리',
-    link: '/resumerry',
+    name: '블로그',
+    link: '/blog',
   },
 ]
