+++ app/mentoring/page.tsx
@@ -1,0 +1,7 @@
+export default function Mentoring() {
+  return (
+    <div>
+      <h4>멘토링</h4>
+    </div>
+  )
+}
